Add signal::windows namespace and helper functions
To implement firwin, there is a need to get_windows on all window types
defined in the Scipy signal.windows namespace, most of which are
implemented in a _windows.py file. This commit keeps all the window
variants that needs to be implemented as comments in an Enum for future
implementation, along with some functions that are used across the
implementation of all window types.

The intended strategy to combat the variadic function type associated to
the Scipy's get_windows function is to enclose all function arguments in
a struct that represents the given function type, while ensuring all
structs representing the Windows type implement a common trait.