Implement path that handles the unexpected shape of zi
ShapeBuilder trait along with Array::as_slice allows as to replicate the
stride_tricks method to create a different array view. It is not yet
presently clear if as_slice_memory_order should be used instead, but the
tests that are included are passing.

It might be possible to remove the reallocation of zi at the start of
the scope now that this is settled.