Add axis_slice function with bad signature
The functionality currently works, but the signature cannot explicitly
specify the return type to be `D`. This is made possible by using heap
allocation of the `SliceInfo` through `Vec` instead of using the
macro/trait method in Lfilter, in an attempt have another method to
circumvent the need for a nightly compiler that uses similar trait
bounds as NDArray without needing a nightly compiler.

This function is eventually used in even/odd/const extensions (which has
only been implemented on nalgebra arrays and not NDArrays) so far.