Fix axis_slice for larger indices
Previously, it was assumed that the only valid values of start and end
must be contained within `0 <= idx < x.shape()[axis]` after unwrapping
the negative and counting from the back (starting with -1), hence the
use of a max clamp. However, this was wrong, as can be validated against
scipy internal.
Thus, a case for this "more negative" than `axis_len` is spelt out, with
the appropriate fix (removing clamp to 0) is done.